Add host builder extensions for package usage
